tests/wpt/tests/svg/painting/parsing/stroke-linejoin-computed.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: getComputedStyle().strokeLinejoin</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#StrokeLinejoinProperty"/>
    <h:meta name="assert" content="stroke-linejoin computed value is as specified."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/computed-testcommon.js"/>
  <script><![CDATA[

test_computed_value("stroke-linejoin", "miter");
test_computed_value("stroke-linejoin", "round");
test_computed_value("stroke-linejoin", "bevel");
// "miter-clip" and "arcs" are not yet supported by browsers.

  ]]></script>
</svg>
